requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
coverage==7.4.2
crispy-bootstrap5==2023.10
Django==4.2.8
django-auth-ldap==4.2.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-log-viewer==1.1.7
django-tables2==2.6.0
dxpy==0.350.1
gunicorn==20.1.0
pandas==2.0.3
plotly==5.15.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dateutil==2.9.0
python-dotenv==1.0.0
python-ldap==3.4.3
pytz==2022.7.1
regex==2023.6.3
sqlparse==0.4.3